Listado de las altas y cambios varios

+++ src/main/java/interfacesDAO/PublicadorDAO.java
@@ -9,6 +9,6 @@ import clases.Publicador;
 
 public interface PublicadorDAO extends GenericDAO<Publicador> {
 	List<Publicacion> obtenerPublicaciones(long id);
-	List<Cartelera> obtenerCarteleras(long id);
+	List<Cartelera> obtenerCarteleras(int id);
 	List<Comentario> obtenerComentarios(long id);
 }
